[FEAT] [RACE DETECTOR] Static track: gpu.barrier is a fence position under fence order
Stage 1b of the option-A plan (design-fence-order.md): the TTIR reader
records tl.debug_barrier's lowering (gpu.barrier) as a fence position
between the surrounding accesses (a half-integer program_seq, since the
encoder numbers accesses densely), the encoder carries the positions on
GlobalEncoding.fence_seqs, and the compiled client passes them with the
fence_order switch into every two-copy solver it builds (T1, the launch
rung, T0 groups). So the static track follows TRITON_VIZ_FENCE_ORDER
exactly like the dynamic one; off keeps every existing verdict.

Tests: four static-track litmus tests join tests/end_to_end/
test_fence_order.py (12 total): the reader records the fence at 0.5,
the unfenced store-then-load races, the fenced one is race-free, and
the legacy reading is unchanged. The compiled-track e2e suite passes
with the switch off (42 passed).

+++ triton_viz/clients/race_detector/two_copy_symbolic_hb_solver.py
@@ -307,7 +307,7 @@ class TwoCopySymbolicHBSolver:
         only_pairs: frozenset[tuple[int, int]] | None = None,
         enum_fallback_grid: tuple[int, int, int] | None = None,
         launch_ceiling: bool = False,
-        fence_seqs: tuple[int, ...] = (),
+        fence_seqs: tuple[float, ...] = (),
         fence_order: bool = False,
     ) -> None:
         self.records = list(records)
@@ -319,7 +319,7 @@ class TwoCopySymbolicHBSolver:
         # between them, and same-instance pairs of distinct operations that
         # no fence separates enter the intra-instance race query. Off keeps
         # the legacy reading (every earlier operation orders every later one).
-        self.fence_seqs: tuple[int, ...] = tuple(sorted(int(x) for x in fence_seqs))
+        self.fence_seqs: tuple[float, ...] = tuple(sorted(float(x) for x in fence_seqs))
         self.fence_order = bool(fence_order)
         self.grid = self._normalize_grid(grid)
         self.arange_dict = dict(arange_dict or {})
@@ -1025,7 +1025,7 @@ class TwoCopySymbolicHBSolver:
 
     # ──────────────────────── Edges & HB closure ────────────────────────
 
-    def _fence_between(self, seq1: int, seq2: int) -> bool:
+    def _fence_between(self, seq1: float, seq2: float) -> bool:
         """True when a captured tile-level fence lies strictly between the
         two program_seq positions (both orientations accepted)."""
         lo, hi = (seq1, seq2) if seq1 <= seq2 else (seq2, seq1)
